+++ clisops/utils/dataset_utils.py
@@ -1,0 +1,71 @@
+import numpy as np
+from roocs_utils.xarray_utils.xarray_utils import get_coord_by_type
+
+from clisops import logging
+
+LOGGER = logging.getLogger(__file__)
+
+
+def calculate_offset(lon, first_element_value):
+    """
+    Calculate the number of elements to roll the dataset by in order to have
+    longitude from -180 to 180 degrees.
+
+    :param lon: longitude coordinate of xarray dataset.
+    :param first_element_value: the value of the first element of the longitude array to roll to.
+    """
+    # get resolution of data
+    res = lon.values[1] - lon.values[0]
+
+    # calculate how many degrees to move by to have lon[0] of rolled subset as lower bound of request
+    diff = first_element_value - lon.values[0]
+
+    # work out how many elements to roll by to roll data by 1 degree
+    index = 1 / res
+
+    # calculate the corresponding offset needed to change data by diff
+    offset = int(diff * index)
+
+    return diff, offset
+
+
+def check_lon_alignment(ds, lon_bnds):
+    """
+    Check whether the longitude subset requested is within the bounds of the dataset. If not try to roll the dataset so
+    that the request is. Raise an exception if rolling is not possible.
+    """
+    low, high = lon_bnds
+    lon = get_coord_by_type(ds, "longitude", ignore_aux_coords=False)
+    lon = ds.coords[lon.name]
+    lon_min, lon_max = lon.values.min(), lon.values.max()
+
+    # handle the case where there is only one longitude
+    if len(lon.values) == 1:
+        lon_value = ds.lon.values[0]
+        if low > lon_value:
+            ds.coords[lon.name] = ds.coords[lon.name] + 360
+        elif high < lon_value:
+            ds.coords[lon.name] = ds.coords[lon.name] - 360
+        return ds
+
+    # check if the request is in bounds - return ds if it is
+    if (lon_min <= low or np.isclose(low, lon_min, atol=0.5)) and (
+        lon_max >= high or np.isclose(high, lon_max, atol=0.5)
+    ):
+        return ds
+
+    else:
+        # check if lon is a dimension
+        if lon.name not in ds.dims:
+            raise Exception(
+                f"The longitude of this dataset runs from {lon_min:.2f} to {lon_max:.2f}, "
+                f"and rolling could not be completed successfully. "
+                f"Please re-run your request with longitudes between these bounds."
+            )
+        # roll the dataset and reassign the longitude values
+        else:
+            first_element_value = low
+            diff, offset = calculate_offset(lon, first_element_value)
+            ds_roll = ds.roll(shifts={f"{lon.name}": offset}, roll_coords=False)
+            ds_roll.coords[lon.name] = ds_roll.coords[lon.name] + diff
+            return ds_roll