+++ tests/ops/test_subset.py
@@ -388,7 +388,7 @@ def test_area_within_area_subset(load_esgf_test_data):
 
 
 def test_area_within_area_subset_cmip6(load_esgf_test_data):
-    area = (100.0, 10.0, 300.0, 90.0)
+    area = (20.0, 10.0, 250.0, 90.0)
 
     outputs = subset(
         ds=CMIP6_RLDS,
@@ -411,7 +411,7 @@ def test_subset_with_lat_lon_single_values(load_esgf_test_data):
     are not being reversed by the `_check_desc_coords` function in
     `clisops.core.subset`.
     """
-    area = (100.0, 10.0, 300.0, 90.0)
+    area = (20.0, 10.0, 250.0, 90.0)
 
     outputs = subset(
         ds=CMIP6_RLDS,
@@ -615,3 +615,158 @@ def test_time_invariant_subset_with_time(load_esgf_test_data):
             output_type="xarray",
         )
     assert str(exc.value) == "'Dataset' object has no attribute 'time'"
+
+
+# test known bug
+@pytest.mark.skipif(os.path.isdir("/badc") is False, reason="data not available")
+@pytest.mark.skip(reason="bug no longer exists")
+def test_cross_prime_meridian(tmpdir):
+    ds = _load_ds(
+        "/badc/cmip6/data/CMIP6/ScenarioMIP/MIROC/MIROC6/ssp119/r1i1p1f1/day/tas/gn/v20191016"
+        "/tas_day_MIROC6_ssp119_r1i1p1f1_gn_20150101-20241231.nc"
+    )
+
+    with pytest.raises(NotImplementedError) as exc:
+        subset(
+            ds=ds,
+            area=(-5, 50, 30, 65),
+            output_dir=tmpdir,
+            output_type="nc",
+            file_namer="simple",
+        )
+    assert (
+        str(exc.value)
+        == "Input longitude bounds ([-5. 30.]) cross the 0 degree meridian "
+        "but dataset longitudes are all positive."
+    )
+
+
+# test it works when not crossing 0 meridian
+@pytest.mark.skipif(os.path.isdir("/badc") is False, reason="data not available")
+def test_do_not_cross_prime_meridian(tmpdir):
+    ds = _load_ds(
+        "/badc/cmip6/data/CMIP6/ScenarioMIP/MIROC/MIROC6/ssp119/r1i1p1f1/day/tas/gn/v20191016"
+        "/tas_day_MIROC6_ssp119_r1i1p1f1_gn_20150101-20241231.nc"
+    )
+
+    result = subset(
+        ds=ds,
+        area=(10, 50, 30, 65),
+        output_dir=tmpdir,
+        output_type="nc",
+        file_namer="simple",
+    )
+
+    _check_output_nc(result)
+
+
+@pytest.mark.skipif(os.path.isdir("/badc") is False, reason="data not available")
+def test_0_360_no_cross(tmpdir):
+
+    ds = _load_ds(
+        "/badc/cmip6/data/CMIP6/CMIP/IPSL/IPSL-CM6A-LR/historical/r1i1p1f1/Amon/rlds/gr/v20180803"
+        "/rlds_Amon_IPSL-CM6A-LR_historical_r1i1p1f1_gr_185001-201412.nc"
+    )
+    result = subset(
+        ds=ds,
+        area=(10.0, -90.0, 200.0, 90.0),
+        output_dir=tmpdir,
+        output_type="nc",
+        file_namer="simple",
+    )
+
+    _check_output_nc(result)
+
+
+@pytest.mark.skipif(os.path.isdir("/badc") is False, reason="data not available")
+@pytest.mark.skip(reason="bug no longer exists")
+def test_0_360_cross(tmpdir):
+    ds = _load_ds(
+        "/badc/cmip6/data/CMIP6/CMIP/IPSL/IPSL-CM6A-LR/historical/r1i1p1f1/Amon/rlds/gr/v20180803/"
+        "rlds_Amon_IPSL-CM6A-LR_historical_r1i1p1f1_gr_185001-201412.nc"
+    )
+
+    with pytest.raises(NotImplementedError):
+        subset(
+            ds=ds,
+            area=(-50.0, -90.0, 100.0, 90.0),
+            output_dir=tmpdir,
+            output_type="nc",
+            file_namer="simple",
+        )
+
+
+@pytest.mark.skipif(os.path.isdir("/badc") is False, reason="data not available")
+def test_300_60_no_cross(tmpdir):
+    # longitude is -300 to 60
+    ds = _load_ds(
+        "/badc/cmip6/data/CMIP6/CMIP/NOAA-GFDL/GFDL-ESM4/historical/r1i1p1f1/Ofx/areacello/gn/v20190726/*.nc"
+    )
+
+    result = subset(
+        ds=ds,
+        area=(10.0, -90.0, 50.0, 90.0),
+        output_dir=tmpdir,
+        output_type="nc",
+        file_namer="simple",
+    )
+
+    _check_output_nc(result)
+
+
+@pytest.mark.skipif(os.path.isdir("/badc") is False, reason="data not available")
+def test_300_60_cross(tmpdir):
+    # longitude is -300 to 60
+    ds = _load_ds(
+        "/badc/cmip6/data/CMIP6/CMIP/NOAA-GFDL/GFDL-ESM4/historical/r1i1p1f1/Ofx/areacello/gn/v20190726/*.nc"
+    )
+
+    result = subset(
+        ds=ds,
+        area=(-100.0, -90.0, 50.0, 90.0),
+        output_dir=tmpdir,
+        output_type="nc",
+        file_namer="simple",
+    )
+
+    _check_output_nc(result)
+
+
+@pytest.mark.skipif(os.path.isdir("/badc") is False, reason="data not available")
+def test_check_lon_alignment_rolled():
+    ds = _load_ds(
+        "/badc/cmip6/data/CMIP6/CMIP/IPSL/IPSL-CM6A-LR/historical/r1i1p1f1/Amon/rlds/gr/v20180803/"
+        "rlds_Amon_IPSL-CM6A-LR_historical_r1i1p1f1_gr_185001-201412.nc"
+    )
+
+    area = (-50.0, -90.0, 100.0, 90.0)
+
+    result = subset(
+        ds=ds,
+        area=area,
+        output_type="xarray",
+    )
+
+    assert area[0] <= all(result[0].lon.data) <= area[2]
+    assert area[1] <= all(result[0].lat.data) <= area[3]
+
+
+@pytest.mark.skipif(os.path.isdir("/badc") is False, reason="data not available")
+def test_check_lon_alignment_irregular_grid():
+    ds = _load_ds(
+        "/badc/cmip6/data/CMIP6/ScenarioMIP/NCC/NorESM2-MM/ssp370/r1i1p1f1/Ofx/sftof/gn/v20191108/*.nc"
+    )
+
+    area = (-50.0, -90.0, 100.0, 90.0)
+
+    with pytest.raises(Exception) as exc:
+        subset(
+            ds=ds,
+            area=area,
+            output_type="xarray",
+        )
+    assert (
+        str(exc.value) == "The longitude of this dataset runs from 0.00 to 359.99, "
+        "and rolling could not be completed successfully. "
+        "Please re-run your request with longitudes between these bounds."
+    )